Fix various type-hints
